refactor(retrieval): merge the predictive-analysis parallel stack into the package
annOptimizer's five ANN variants become named, documented modules under
algorithms/vector/strategies/ (exact scan, ivf, prefiltered, matryoshka,
adaptive hybrid), with ANNOptimizer surviving as the config-driven
dispatcher — documentQ&A's AIQuery/AIChat routes now import it from the
package instead of reaching into another feature's folder. The feature's
private hybridSearch dissolves along the design's lines: the ts_rank
full-text leg into algorithms/bm25/fts.ts, generic Reciprocal Rank Fusion
plus the composed page-level hybrid search into algorithms/fusion/, with
the query embedder injected rather than chosen by the algorithm.

Two of the three independent embedding clients are retired: predictive's
utils/embeddings and all three notes-vertical OpenAIEmbeddings
instantiations now generate through @launchstack/llm's embedding service
(same endpoints, models, and dimensions — the files keep only env-pair
resolution, caching, and error semantics). One embedding transport remains.

Verified: workspace typecheck, lint at baseline, retrieval + tools vitest,
full apps/web jest (2242 passed) against the migrated pgvector database.
