+++ apps/web/src/app/api/agents/documentQ&A/AIChat/query/route.ts
@@ -2,13 +2,12 @@ import { NextResponse } from "next/server";
 import { SystemMessage, HumanMessage } from "@langchain/core/messages";
 import { db } from "~/server/db/index";
 import { and, eq, inArray } from "drizzle-orm";
-import ANNOptimizer from "~/app/api/agents/predictive-document-analysis/services/annOptimizer";
+import { ANNOptimizer, createDocumentVectorRetriever } from "@launchstack/retrieval/algorithms/vector";
 import {
     companyEnsembleSearch,
     documentEnsembleSearch,
     multiDocEnsembleSearch,
 } from "~/server/rag/ensemble";
-import { createDocumentVectorRetriever } from "@launchstack/retrieval/algorithms/vector";
 import type {
     CompanySearchOptions,
     DocumentSearchOptions,
