+++ apps/web/src/app/api/agents/predictive-document-analysis/utils/embeddings.ts
@@ -1,4 +1,15 @@
-import { OpenAIEmbeddings } from "@langchain/openai";
+/**
+ * Query embeddings for predictive analysis, generated through
+ * @launchstack/llm's embedding service (no private HTTP client here — this
+ * file only resolves the feature's endpoint pair, caches, and sets error
+ * semantics). Fixed at text-embedding-3-large / 1536 dims because every
+ * similarity comparison this feature makes is against the legacy 1536-dim
+ * `documentSections` / `documentRetrievalChunks` embeddings.
+ */
+
+import { generateEmbeddings } from "@launchstack/llm/embeddings";
+import { LRUCache } from "lru-cache";
+import { sanitizeErrorMessage } from "~/app/api/agents/predictive-document-analysis/utils/logging";
 
 /**
  * Thrown when embeddings are not configured at all. Distinct from a transient
@@ -13,18 +24,17 @@ export class EmbeddingConfigurationError extends Error {
 /**
  * Endpoint and credential, resolved as a PAIR.
  *
- * `@langchain/openai` silently falls back to `api.openai.com` when
- * `configuration.baseURL` is undefined, so a key without an endpoint used to
- * ship document text to a vendor nothing here names. Both halves come from the
- * same source or the call fails.
+ * The llm embedding service deliberately has no default endpoint —
+ * embeddings are persisted, so the provider must be named explicitly. Both
+ * halves come from the same source or the call fails.
  */
-function resolveEmbeddingEndpoint(): { apiKey: string; baseURL: string } {
-    const baseURL = process.env.EMBEDDING_API_BASE_URL ?? process.env.AI_BASE_URL;
+function resolveEmbeddingEndpoint(): { apiKey: string; baseUrl: string } {
+    const baseUrl = process.env.EMBEDDING_API_BASE_URL ?? process.env.AI_BASE_URL;
     const apiKey = process.env.EMBEDDING_API_BASE_URL
         ? process.env.EMBEDDING_API_KEY
         : (process.env.AI_API_KEY ?? process.env.OPENAI_API_KEY);
 
-    if (!baseURL || !apiKey) {
+    if (!baseUrl || !apiKey) {
         throw new EmbeddingConfigurationError(
             "Embeddings are not configured. Set EMBEDDING_API_BASE_URL and " +
                 "EMBEDDING_API_KEY (or AI_BASE_URL and AI_API_KEY). There is no " +
@@ -32,12 +42,11 @@ function resolveEmbeddingEndpoint(): { apiKey: string; baseURL: string } {
                 "must be named explicitly."
         );
     }
-    return { apiKey, baseURL };
+    return { apiKey, baseUrl };
 }
-import { LRUCache } from "lru-cache";
-import { sanitizeErrorMessage } from "~/app/api/agents/predictive-document-analysis/utils/logging";
 
 const EMBEDDING_MODEL = "text-embedding-3-large";
+const EMBEDDING_DIMENSIONS = 1536;
 const MAX_CACHE_ENTRIES = 500;
 
 const embeddingCache = new LRUCache<string, number[]>({
@@ -51,16 +60,14 @@ export async function getEmbeddings(text: string): Promise<number[]> {
     }
 
     try {
-        const { apiKey, baseURL } = resolveEmbeddingEndpoint();
-        const embeddings = new OpenAIEmbeddings({
-            openAIApiKey: apiKey,
-            modelName: EMBEDDING_MODEL,
-            dimensions: 1536,
-            configuration: { baseURL },
+        const { apiKey, baseUrl } = resolveEmbeddingEndpoint();
+        const { embeddings } = await generateEmbeddings([text], {
+            apiKey,
+            baseUrl,
+            model: EMBEDDING_MODEL,
+            dimensions: EMBEDDING_DIMENSIONS,
         });
-
-        const [embedding] = await embeddings.embedDocuments([text]);
-        const result = embedding ?? [];
+        const result = embeddings[0] ?? [];
 
         embeddingCache.set(text, result);
         return result;
@@ -78,16 +85,14 @@ export async function batchGetEmbeddings(texts: string[]): Promise<number[][]> {
     const uniqueTexts = [...new Set(texts)];
 
     try {
-        const { apiKey, baseURL } = resolveEmbeddingEndpoint();
-        const embeddings = new OpenAIEmbeddings({
-            openAIApiKey: apiKey,
-            modelName: EMBEDDING_MODEL,
-            dimensions: 1536,
-            configuration: { baseURL },
+        const { apiKey, baseUrl } = resolveEmbeddingEndpoint();
+        const { embeddings } = await generateEmbeddings(uniqueTexts, {
+            apiKey,
+            baseUrl,
+            model: EMBEDDING_MODEL,
+            dimensions: EMBEDDING_DIMENSIONS,
         });
-
-        const results = await embeddings.embedDocuments(uniqueTexts);
-        const embeddingMap = new Map(uniqueTexts.map((text, i) => [text, results[i]]));
+        const embeddingMap = new Map(uniqueTexts.map((text, i) => [text, embeddings[i]]));
 
         embeddingMap.forEach((embedding, text) => {
             embeddingCache.set(text, embedding ?? []);
