+++ packages/retrieval/src/algorithms/bm25/index.ts
@@ -7,3 +7,5 @@ export {
     createCompanyBM25Retriever,
     createMultiDocBM25Retriever,
 } from "./bm25";
+
+export { ftsSearch } from "./fts";
