+++ packages/retrieval/src/algorithms/fusion/index.ts
@@ -1,0 +1,2 @@
+export { reciprocalRankFusion, type RankedResult, type FusedResult } from "./rrf";
+export { hybridSearchWithRRF, type HybridSearchMatch } from "./hybrid-search";
