+++ packages/retrieval/src/algorithms/vector/index.ts
@@ -6,3 +6,14 @@ export {
 } from "./retriever";
 
 export { cosineSimilarity, euclideanDistance } from "./similarity";
+
+export {
+    ANNOptimizer,
+    exactScanSearch,
+    ivfSearch,
+    prefilteredSearch,
+    matryoshkaSearch,
+    buildDocumentCluster,
+    calculateDocumentRelevanceScores,
+    findRelevantDocumentClusters,
+} from "./strategies";
